+++ resources/aws_support_app_slack_channel_configuration.rb
@@ -1,7 +1,6 @@
 # Import API specifics
 use "awscc_base"
 
-unified_mode true
 resource_name :aws_support_app_slack_channel_configuration
 provides :aws_support_app_slack_channel_configuration, target_mode: true, platform: "aws"